minor changes to Setting structs
- add extra header for setting structs
- add find method to every struct
